- Added explicit creation of indices in CirculantSparseMatrix as int32. - Added "plot" method for each experiment object. - Added Heat models, both in Forward-Euler method and in Leap-Frog method. - Fixing Leap-Frog class for adjusting (needs further attention) for nonhomogeneous equations (2 * ). - Added a nonhomogeneous term, to be used with the heat models.

+++ Infrastructure/circulant_sparse_matrix.py
@@ -7,7 +7,7 @@ class CirculantSparseMatrix(object):
     def __init__(self, n, nonzero_terms, nonzero_indices):
         self._n = n
         self._terms = np.array(nonzero_terms)
-        self._indices = np.array(nonzero_indices)
+        self._indices = np.array(nonzero_indices, dtype=np.int32)
 
     def dot(self, v):
         next_state = np.zeros(v.shape)
